added migrations

+++ api/Dtos/Stock.cs
@@ -4,7 +4,7 @@
 namespace api.Dtos.Stock{
     public class StockDto{
         public int Id { get; set; }
-        public int Purchase { get; set; }
+        public decimal Purchase { get; set; }
         public string CompanyName { get; set; } = string.Empty;
         public string Industry { get; set; } = string.Empty;
         public string Ticker { get; set; } = string.Empty;
