+++ api/Models/Stock.cs
@@ -13,7 +13,7 @@ namespace api.Models{
         [Key]
         public int Id { get; set; }
         [Column(TypeName = "decimal(12,2)")]
-        public int Purchase { get; set; }
+        public decimal Purchase { get; set; }
         public string CompanyName { get; set; } = string.Empty;
         public string Industry { get; set; } = string.Empty;
         public string Ticker { get; set; } = string.Empty;
